--- src/constants/index.js
@@ -1,4 +1,0 @@
-export * from './colors';
-export * from './data';  
-
-
