+++ src/index.css
@@ -28,4 +28,6 @@
     .projects-grid {
         @apply grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 max-w-7xl mx-auto px-4 sm:px-6 lg:px-8;
     }
+
+
 }
